refactor(platform): move ProcessAlive into platform/utils
The signal-0 liveness probe sat in internal/services with two consumers -
cmd/daemon.go and the gateway manager's PID registry. It is a stdlib wrapper
with no service in it; platform/utils is where it belongs.

Drop the process_alive.go wrapper and export ProcessAlive directly from the
two build-tagged implementations rather than carrying a one-line indirection.
The //go:build tags travel with the files.

Adds the test the issue asked for: self is alive, a spawned sleep is alive,
and the same PID is dead once killed and reaped.
