+++ internal/platform/utils/process_alive_other.go
@@ -1,12 +1,12 @@
 //go:build !unix
 
-package services
+package utils
 
 import "os"
 
-// processAlive reports whether a process with the given PID exists. On Windows
+// ProcessAlive reports whether a process with the given PID exists. On Windows
 // os.FindProcess opens a handle to the process and fails if it does not exist.
-func processAlive(pid int) bool {
+func ProcessAlive(pid int) bool {
 	p, err := os.FindProcess(pid)
 	if err != nil {
 		return false
