--- internal/services/process_alive.go
@@ -1,6 +1,0 @@
-package services
-
-// ProcessAlive reports whether a process with the given PID exists.
-func ProcessAlive(pid int) bool {
-	return processAlive(pid)
-}
